+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [go v0.21.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [Go v0.21.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.21.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-10-16 01:24:54
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.21.1] previous-versions/sync-for-commerce-version-1
+### Releases
+- [Go v0.21.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.21.1 - previous-versions/sync-for-commerce-version-1